learning_journey/Day3_ActionChain_Agent/requirements.txt
langchain==0.2.11
langchain-community==0.2.7
langchain-core==0.2.26
python-dotenv==1.0.1
requests